Rename module
Rename option_editor.py from OptionEditor.py to comply with python file naming convention
